requirements.txt
streamlit
azure-ai-formrecognizer
azure-core
python-dotenv
pandas
sentence-transformers
pymongo
